+++ library/src/compute_scheme.cpp
@@ -58,7 +58,6 @@ static const std::map<ComputeScheme, const char*>& ComputeSchemetoStringMap()
            {ENUMSTR(CS_KERNEL_TRANSPOSE_CMPLX_TO_R)},
            {ENUMSTR(CS_REAL_2D_EVEN)},
            {ENUMSTR(CS_REAL_3D_EVEN)},
-           {ENUMSTR(CS_KERNEL_APPLY_CALLBACK)},
 
            {ENUMSTR(CS_BLUESTEIN)},
            {ENUMSTR(CS_KERNEL_CHIRP)},