+++ library/src/rtc_kernel.cpp
@@ -146,8 +146,6 @@ std::shared_future<std::unique_ptr<RTCKernel>>
         generator = RTCKernelBluesteinSingle::generate_from_node(node, gpu_arch, enable_callbacks);
     if(!generator.valid())
         generator = RTCKernelBluesteinMulti::generate_from_node(node, gpu_arch, enable_callbacks);
-    if(!generator.valid())
-        generator = RTCKernelApplyCallback::generate_from_node(node, gpu_arch, enable_callbacks);
     if(generator.valid())
     {
         kernel_name = generator.generate_name();